feat: implement hoursInRadians with basic π return to make test pass
